zha_button_click: refactored for high performance
I completely rewrote this automation for counting button clicks as quick and efficient as possible. Therefore the first action is to increment the click counter (using a counter variable this time). Before the rest of the processing happens.

1.  Increment the counter then delay 0.5 seconds.
2. If the action has not be canceled duing the previous 1/2 second, then store the click count and reset the counter.
3. Then make sense of the event data.
4. Pass the action portion off to `script.zha_button_actions`.

I believe this makes the entire process work more reliably, and overall the code is easier to understand. I am thrilled with how this is turning out.